requirements.txt
discord.py==2.3.1
discord-pretty-help==2.0.7
discord.py_pagination==0.0.5
gspread==5.10.0
Paginator==0.5.1
python-dotenv==1.0.0
